finished signin and signup modal form and allows switching in the modal
