--- frontend/src/components/LoginFormPage/index.js
@@ -1,3 +1,0 @@
-import LoginFormPage from './LoginFormPage';
-
-export default LoginFormPage;
