--- frontend/src/components/SignupFormPage/index.js
@@ -1,3 +1,0 @@
-import SignupFormPage from './SignupFormPage';
-
-export default SignupFormPage;
